+++ tools/progen/TabString.h
@@ -7,13 +7,14 @@
 
 #include <map>
 #include <stddef.h>
+#include <stdint.h>
 
 class TabString
 {
-	typedef std::map<size_t, const char*> TabStringMap;
+	typedef std::map<uint32_t, const char*> TabStringMap;
 public:
 	~TabString();
-	static const char* get(size_t tabCount);
+	static const char* get(uint32_t tabCount);
 private:
 	TabString();
 	static TabString& instance();
